refactor: Home 컴포넌트 styled 코드 분리 및 props 코드 리팩토링
